Move app/config.js to particle.app.config.js
