+++ apps/grav/webpack.config.js
@@ -14,7 +14,7 @@ const { NODE_ENV } = process.env;
 // Constants: root
 const { ASSETS_ATOMIC_FOLDER } = require('../../config');
 // Constants: app
-const appConfig = require('./config');
+const appConfig = require('./particle.app.config');
 
 const { APP_NAME, APP_DESIGN_SYSTEM, APP_DIST, APP_DIST_PUBLIC } = appConfig;
